Added geometry rendering, might remove

+++ src/vector2.rs
@@ -3,6 +3,6 @@ use serde::Serialize;
 
 #[derive(Copy, Clone, Debug, Deserialize, Serialize)]
 pub struct Vector2 {
-    pub x: f64,
-    pub y: f64,
+    pub x: f32,
+    pub y: f32,
 }
